+++ js/src/main/scala/zio/pdf/PdfEngine.scala
@@ -494,7 +494,7 @@ object PdfEngine:
 
   def watermark(
     bytes: Chunk[Byte],
-    stamp: PdfWatermark.Text,
+    stamp: PdfWatermark.Stamp,
     opts: Options = Options.default
   ): ZIO[PdfEngine, Throwable, Chunk[Byte]] =
     PdfWatermark.fromBytes(bytes, stamp, opts)